Safari-609.1.11.1/LayoutTests/storage/domstorage/localstorage/set-value-in-storage-after-window-close-expected.txt

Tests that we can't store a value in local storage after calling window.close().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.localStorage is defined.
PASS window.localStorage["test"] is undefined.
PASS successfullyParsed is true

TEST COMPLETE
